Better testing for the library
This commit ensures that each feature of the library is
tested. Everything is done using the 'default' scenario
and 'create', 'prepare' and 'converge' sequences.
Ansible tags are used to determine which tests must be
run and usage of 'set_fact' module with fact caching
ensure that the same test can be used for each feature.