Ensure the partition reassignment is finished (thus the reassignement znode is missing) after updating the number of replica.
Add a new method to ensure the znode is missing before AND after updating the number of replica for a topic.
Add a new method to refresh topics state after an action was done.
Add an example (and a warning) in the README.md file.
Add a test to verify the behaviour.

Fixes https://github.com/StephenSorriaux/ansible-kafka-admin/issues/46